% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta2norm}
\alias{beta2norm}
\title{Convert Beta to Normal}
\usage{
beta2norm(x, shape1 = 1, shape2 = 1)
}
\arguments{
\item{x}{Numeric vector of beta-distributed values.}

\item{shape1, shape2}{Beta distribution shape parameters.}
}
\value{
Numeric vector of normal-distributed values.
}
\description{
Convert Beta to Normal
}
