% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binom2norm}
\alias{binom2norm}
\title{Convert Binomial to Normal}
\usage{
binom2norm(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{Integer vector of binomial counts.}

\item{size}{Number of trials.}

\item{prob}{Probability of success.}
}
\value{
Numeric vector.
}
\description{
Convert Binomial to Normal
}
