% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_anova}
\alias{contr_code_anova}
\title{ANOVA (sum) Contrast Matrix — alias for contr_code_sum}
\usage{
contr_code_anova(k)
}
\arguments{
\item{k}{Integer. Number of levels.}
}
\description{
ANOVA (sum) Contrast Matrix — alias for contr_code_sum
}
