% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_difference}
\alias{contr_code_difference}
\title{Difference (Sequential) Contrast Matrix}
\usage{
contr_code_difference(k)
}
\arguments{
\item{k}{Integer. Number of levels.}
}
\value{
A \verb{k x (k-1)} matrix.
}
\description{
Each column contrasts adjacent level pairs: level i+1 vs level i.
}
