% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_helmert}
\alias{contr_code_helmert}
\title{Helmert Contrast Matrix}
\usage{
contr_code_helmert(k)
}
\arguments{
\item{k}{Integer. Number of levels.}
}
\value{
A \verb{k x (k-1)} matrix.
}
\description{
Helmert Contrast Matrix
}
