% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_poly}
\alias{contr_code_poly}
\title{Polynomial Contrast Matrix}
\usage{
contr_code_poly(k)
}
\arguments{
\item{k}{Integer. Number of levels.}
}
\value{
A \verb{k x (k-1)} matrix.
}
\description{
Polynomial Contrast Matrix
}
