% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_sum}
\alias{contr_code_sum}
\title{Sum (Deviation) Contrast Matrix}
\usage{
contr_code_sum(k)
}
\arguments{
\item{k}{Integer. Number of levels.}
}
\value{
A \verb{k x (k-1)} matrix.
}
\description{
Sum (Deviation) Contrast Matrix
}
