% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contr_code_treatment}
\alias{contr_code_treatment}
\title{Treatment Contrast Matrix}
\usage{
contr_code_treatment(k, reference = 1)
}
\arguments{
\item{k}{Integer. Number of levels.}

\item{reference}{Integer. Reference level index.}
}
\value{
A \verb{k x (k-1)} matrix.
}
\description{
Treatment Contrast Matrix
}
