% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma2norm}
\alias{gamma2norm}
\title{Convert Gamma to Normal}
\usage{
gamma2norm(x, shape = 1, rate = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{shape}{Shape parameter.}

\item{rate}{Rate parameter.}
}
\value{
Numeric vector.
}
\description{
Convert Gamma to Normal
}
