% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{is_pos_def}
\alias{is_pos_def}
\title{Check Whether a Matrix is Positive Definite}
\usage{
is_pos_def(m)
}
\arguments{
\item{m}{A square matrix.}
}
\value{
Logical scalar.
}
\description{
Check Whether a Matrix is Positive Definite
}
