% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{likert2norm}
\alias{likert2norm}
\title{Convert Likert to Normal}
\usage{
likert2norm(x, items = 5)
}
\arguments{
\item{x}{Integer vector of Likert responses.}

\item{items}{Integer. Number of scale points.}
}
\value{
Numeric vector of approximately normal values.
}
\description{
Convert Likert to Normal
}
