% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2beta}
\alias{norm2beta}
\title{Convert Normal to Beta}
\usage{
norm2beta(x, shape1 = 1, shape2 = 1)
}
\arguments{
\item{x}{Numeric vector (standard normal or from \code{\link{rnorm_forge}}).}

\item{shape1, shape2}{Beta distribution shape parameters.}
}
\value{
Numeric vector of beta-distributed values.
}
\description{
Convert Normal to Beta
}
\examples{
norm2beta(rnorm(100), shape1 = 2, shape2 = 5)
}
