% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2binom}
\alias{norm2binom}
\title{Convert Normal to Binomial}
\usage{
norm2binom(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{Numeric vector.}

\item{size}{Number of trials.}

\item{prob}{Probability of success.}
}
\value{
Integer vector.
}
\description{
Convert Normal to Binomial
}
\examples{
norm2binom(rnorm(100), size = 10, prob = 0.3)
}
