% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2gamma}
\alias{norm2gamma}
\title{Convert Normal to Gamma}
\usage{
norm2gamma(x, shape = 1, rate = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{shape}{Shape parameter.}

\item{rate}{Rate parameter (1/scale).}
}
\value{
Numeric vector.
}
\description{
Convert Normal to Gamma
}
\examples{
norm2gamma(rnorm(100), shape = 2, rate = 1)
}
