% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2likert}
\alias{norm2likert}
\title{Convert Normal to Likert Scale}
\usage{
norm2likert(x, items = 5)
}
\arguments{
\item{x}{Numeric vector (latent normal values).}

\item{items}{Integer. Number of scale points.}
}
\value{
Integer vector with values in \code{1:items}.
}
\description{
Convert Normal to Likert Scale
}
