% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2pois}
\alias{norm2pois}
\title{Convert Normal to Poisson}
\usage{
norm2pois(x, lambda = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{lambda}{Mean of the Poisson distribution.}
}
\value{
Integer vector.
}
\description{
Convert Normal to Poisson
}
\examples{
norm2pois(rnorm(100), lambda = 3)
}
