% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2trunc}
\alias{norm2trunc}
\title{Convert Normal to Truncated Normal}
\usage{
norm2trunc(x, min = -Inf, max = Inf, mu = 0, sd = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{min, max}{Truncation bounds. \code{NA} means no truncation on that side.}

\item{mu}{Mean of the (untruncated) normal.}

\item{sd}{SD of the (untruncated) normal.}
}
\value{
Numeric vector truncated to \code{min} and \code{max}.
}
\description{
Convert Normal to Truncated Normal
}
\examples{
norm2trunc(rnorm(200), min = 0, max = 5, mu = 2, sd = 1.5)
}
