% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm2unif}
\alias{norm2unif}
\title{Convert Normal to Uniform}
\usage{
norm2unif(x, min = 0, max = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{min}{Lower bound.}

\item{max}{Upper bound.}
}
\value{
Numeric vector.
}
\description{
Convert Normal to Uniform
}
\examples{
norm2unif(rnorm(100), min = 0, max = 10)
}
