% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_dist.R
\name{plikert}
\alias{plikert}
\title{Likert Cumulative Distribution Function}
\usage{
plikert(q, items = 5, mu = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Integer vector of Likert values.}

\item{items}{Integer. Number of scale points.}

\item{mu, sd}{Parameters of the latent normal.}

\item{lower.tail}{Logical.}

\item{log.p}{Logical.}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Likert Cumulative Distribution Function
}
