% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pois2norm}
\alias{pois2norm}
\title{Convert Poisson to Normal}
\usage{
pois2norm(x, lambda = 1)
}
\arguments{
\item{x}{Integer vector.}

\item{lambda}{Mean of the Poisson distribution.}
}
\value{
Numeric vector.
}
\description{
Convert Poisson to Normal
}
