% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_dist.R
\name{qlikert}
\alias{qlikert}
\title{Likert Quantile Function}
\usage{
qlikert(p, items = 5, mu = 0, sd = 1)
}
\arguments{
\item{p}{Probability vector.}

\item{items}{Integer. Number of scale points.}

\item{mu, sd}{Parameters of the latent normal.}
}
\value{
Integer vector.
}
\description{
Likert Quantile Function
}
