% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{trunc2norm}
\alias{trunc2norm}
\title{Convert Truncated Normal to Normal}
\usage{
trunc2norm(x, min = -Inf, max = Inf, mu = 0, sd = 1)
}
\arguments{
\item{x}{Numeric vector (truncated normal values).}

\item{min, max}{The bounds that were used to truncate.}

\item{mu, sd}{Parameters of the underlying normal.}
}
\value{
Numeric vector on the standard normal scale.
}
\description{
Convert Truncated Normal to Normal
}
