% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{unif2norm}
\alias{unif2norm}
\title{Convert Uniform to Normal}
\usage{
unif2norm(x, min = 0, max = 1)
}
\arguments{
\item{x}{Numeric vector (uniform values in \code{min} and \code{max}).}

\item{min}{Lower bound.}

\item{max}{Upper bound.}
}
\value{
Numeric vector.
}
\description{
Convert Uniform to Normal
}
